+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\index.jsx":"1","C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\App.jsx":"2","C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\components\\Base.jsx":"3","C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\components\\Home.jsx":"4","C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\templates\\Header\\HeaderStyles.jsx":"5","C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\templates\\Header\\Header.jsx":"6"},{"size":197,"mtime":1608380649523,"results":"7","hashOfConfig":"8"},{"size":419,"mtime":1608383603325,"results":"9","hashOfConfig":"8"},{"size":261,"mtime":1608384293823,"results":"10","hashOfConfig":"8"},{"size":178,"mtime":1608384232046,"results":"11","hashOfConfig":"8"},{"size":0,"mtime":1608379811966,"results":"12","hashOfConfig":"8"},{"size":191,"mtime":1608384458502,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"6n7ijq",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\index.jsx",[],"C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\App.jsx",[],"C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\components\\Base.jsx",[],"C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\components\\Home.jsx",[],"C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\templates\\Header\\HeaderStyles.jsx",[],["27","28"],"C:\\Users\\yuusa\\Desktop\\portfolio\\sisterweb\\src\\templates\\Header\\Header.jsx",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
