+++ package.json
@@ -34,5 +34,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0"
   }
 }
