+++ src/components/Home.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+import Header from "../templates/Header/Header.jsx";
+
+const Home = () => {
+  
+};
+
+
+
+ <Header />
+
+
+
+
+
+
+
+
+
+
+
+
+
+export default Home
