+++ src/templates/Footer/Footer.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import {} from "./FooterStyles";
+
+const Footer = () => {
+  return (
+    <>
+
+    </>
+  );
+};
+
+export default Footer;
