+++ src/templates/Footer/FooterStyles.jsx
@@ -1,0 +1,2 @@
+import styled from "styled-components";
+
