+++ src/templates/Header/Header.jsx
@@ -1,0 +1,23 @@
+import React from "react";
+// import {} from "./HeaderStyles";
+
+
+const Header = () => {
+
+
+return(
+
+<>
+<h1>rhegu</h1>
+<h1>cdewv</h1>
+
+
+
+
+</>
+
+)
+
+}
+
+export default Header;
